Fixed closest and ping event handler.

+++ src/main/kotlin/harpo/network/communication/p2p/infrastructure/messages/type/Ping.kt
@@ -8,22 +8,23 @@ import io.ep2p.kademlia.protocol.message.KademliaMessage
 import io.ep2p.kademlia.protocol.message.PongKademliaMessage
 import java.io.Serializable
 import java.math.BigDecimal
+import java.math.BigInteger
 
 class Ping(private val repository: OperationsRepository): Message {
 
     override fun <INPUT : Serializable?, OUTPUT : Serializable?> sendMessage(
-        self: KademliaNodeAPI<BigDecimal, ConnectionInfoImpl>?,
-        receiver: Node<BigDecimal, ConnectionInfoImpl>?,
-        message: KademliaMessage<BigDecimal, ConnectionInfoImpl, OUTPUT>?
-    ): KademliaMessage<BigDecimal, ConnectionInfoImpl, INPUT> {
+        self: KademliaNodeAPI<BigInteger, ConnectionInfoImpl>?,
+        receiver: Node<BigInteger, ConnectionInfoImpl>?,
+        message: KademliaMessage<BigInteger, ConnectionInfoImpl, OUTPUT>?
+    ): KademliaMessage<BigInteger, ConnectionInfoImpl, INPUT> {
 
         val answer = repository.ping(receiver, self)
 
-        val response = PongKademliaMessage<BigDecimal, ConnectionInfoImpl>()
+        val response = PongKademliaMessage<BigInteger, ConnectionInfoImpl>()
         response?.node = receiver
         response?.isAlive = answer.isAlive
 
-        return response as KademliaMessage<BigDecimal, ConnectionInfoImpl, INPUT>
+        return response as KademliaMessage<BigInteger, ConnectionInfoImpl, INPUT>
     }
 
 }
